Remove pysimplemodel
