+++ pyproject.toml
@@ -26,7 +26,6 @@ PyXB = "1.2.5"
 zeep = "^3.4"
 signxml = "^2.6"
 lxml = "^4.4"
-pysimplemodel = "^2.4"
 ipython = "^7.8"
 pydantic = "^1.0"
